[CLEANUP] Restructured components as classes; renamed ListKingsComponent to KingsComponent; moved KING elements from AccountInfo to King Component.
